Add support for reading credentials from vault
We support both specifying a token and using approles. The new parameters are:

-vaultAddr string
  	Vault address to use. (default "http://localhost:8200")
-vaultRoleId string
  	Role ID to access vault with. Requires vaultSecretId.
-vaultSecretId string
  	Secret ID to access vault with. Requires vaultRoleId.
-vaultToken string
  	Token to access vault with. Takes precedence over vaultRoleId and vaultSecredId.

Leases on renewable tokens are renewed regularly.

When used with an explicit token we expect the secrets to be written to
"secrets/dbwebapp{username:,password:}."
When using app roles "database/creds/vault-mysql-role" is being read to
receive the temporary db credentials.

Both vault and db servers are being pinged to make sure they are up and
running before starting the server, with increasing backoff time.

The parameter precedence works as follows:

1: explicit DBUSER and DBPASS
2: vault token
3: vault approle (roleId and secretId)